+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All changes to the **Hearts of Modding** extension will be documented in this file.
 
+## [v0.21.3] - 2026-06-27
+
+### 🔧 Changed
+
+- **Changed severity of diagnostic `HOM3008` (`MISSING_EVENT_NAMESPACE`) from a warning to an error.** Not declaring a namespace causes `Malformed token` errors and the game fails to register the events.
+
 ## [v0.21.2] - 2026-06-27
 
 ### ✨ Added